expose estimate fees
